Move commands to new folder and namespace
