+++ source/GreenEnergyHub.TimeSeries/source/GreenEnergyHub.TimeSeries.Infrastructure/Messaging/Serialization/Common/DocumentConverterConstants.cs
@@ -31,5 +31,7 @@ namespace GreenEnergyHub.TimeSeries.Infrastructure.Messaging.Serialization.Commo
         internal const string RecipientId = "receiver_MarketParticipant.mRID";
 
         internal const string RecipientBusinessProcessRole = "receiver_MarketParticipant.marketRole.type";
+
+        internal const string CreatedDateTime = "createdDateTime";
     }
 }
